move the utils to utils_helper and remove unuised condition for LinOp as template
